requirements.txt
Flask==3.0.3
gunicorn==22.0.0
azure-data-tables==12.5.0